requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.0
boto3==1.35.47
botocore==1.35.47
cffi==1.17.1
click==8.1.7
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
et-xmlfile==1.1.0
fastapi==0.115.3
fonttools==4.54.1
h11==0.14.0
idna==3.10
jmespath==1.0.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
psycopg2==2.9.10
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
s3transfer==0.10.3
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
XlsxWriter==3.2.0
